pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Task manager"
authors = ["al-ov73 <klever1987@yandex.ru>"]
readme = "README.md"
packages = [
  { include = "task_manager" },
]


[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0"
dj-database-url = "^2.1.0"
psycopg2-binary = "^2.9.9"
whitenoise = {extras = ["brotli"], version = "^6.6.0"}
gunicorn = "^21.2.0"
python-dotenv = "^1.0.0"
django-bootstrap5 = "^23.3"
logging = "^0.4.9.6"
django-filter = "^23.5"
django-extensions = "^3.2.3"
rollbar = "^1.0.0"
flake8 = "^6.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
